+++ frontend/src/main/pages/AdminListCommonPageV2.jsx
@@ -1,12 +1,18 @@
-import React from "react";
+import React, { useState } from "react";
 import BasicLayout from "main/layouts/BasicLayout/BasicLayout";
 import AdminCommonsCard from "main/components/Commons/AdminCommonsCard";
 import { useBackend } from "main/utils/useBackend";
 import { useCurrentUser } from "main/utils/currentUser";
-import { Button, Row, Col, Container } from "react-bootstrap";
+import { Button, Row, Col, Container, Form } from "react-bootstrap";
 
 export default function AdminListCommonsPageV2() {
   const { data: currentUser } = useCurrentUser();
+
+  const [query, setQuery] = useState("");
+
+  const handleQueryChange = (event) => {
+    setQuery(event.target.value);
+  };
 
   // Stryker disable  all
   const {
@@ -20,6 +26,10 @@ export default function AdminListCommonsPageV2() {
   );
   // Stryker restore  all
 
+  const filteredCommons = commons.filter((c) =>
+    c.commons.name.toLowerCase().includes(query.toLowerCase()),
+  );
+
   // Stryker disable all - styles that don't need to be mut tested
   const DownloadButtonStyle = {
     display: "flex",
@@ -30,19 +40,40 @@ export default function AdminListCommonsPageV2() {
   return (
     <BasicLayout>
       <div className="pt-2">
-        <Row className="pt-5">
-          <Col>
-            <h2>Commons</h2>
+        <Row className="pt-5 page-header align-items-center">
+          <Col md={3}>
+            <h2 style={{ margin: 0 }}>Commons</h2>
           </Col>
-          <Col style={DownloadButtonStyle}>
-            <Button href="/api/commonstats/downloadAll">
+          <Col md={6}>
+            <div className="search-container">
+              <Form.Control
+                type="text"
+                placeholder="Search by name..."
+                value={query}
+                onChange={handleQueryChange}
+                data-testid="AdminListCommonsPage-Search"
+                className="search-input"
+              />
+              <span className="search-icon-wrapper">🔍</span>
+            </div>
+          </Col>
+          <Col md={3} style={DownloadButtonStyle}>
+            <Button
+              href="/api/commonstats/downloadAll"
+              variant="outline-success"
+              style={{ borderRadius: "30px", padding: "10px 20px" }}
+            >
               Download All Stats
             </Button>
           </Col>
         </Row>
-        <Container fluid className="pt-3">
-          {commons &&
-            commons.map((common) => (
+        <Container
+          fluid
+          className="pt-3"
+          style={{ animation: "fadeInDown 1s ease-out" }}
+        >
+          {filteredCommons &&
+            filteredCommons.map((common) => (
               <AdminCommonsCard
                 key={common.commons.id}
                 commonItem={common}
