+++ frontend/src/stories/pages/AdminListCommonsPageV2.stories.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import AdminListCommonPage from "main/pages/AdminListCommonPage";
+import AdminListCommonsPageV2 from "main/pages/AdminListCommonPageV2";
 
 import { apiCurrentUserFixtures } from "fixtures/currentUserFixtures";
 import { systemInfoFixtures } from "fixtures/systemInfoFixtures";
@@ -7,11 +7,11 @@ import { http, HttpResponse } from "msw";
 import commonsPlusFixtures from "fixtures/commonsPlusFixtures";
 
 export default {
-  title: "pages/AdminListCommonPage",
-  component: AdminListCommonPage,
+  title: "pages/AdminListCommonsPageV2",
+  component: AdminListCommonsPageV2,
 };
 
-export const adminListPage = () => <AdminListCommonPage />;
+export const adminListPage = () => <AdminListCommonsPageV2 />;
 
 adminListPage.parameters = {
   msw: [
@@ -27,6 +27,28 @@ adminListPage.parameters = {
     }),
     http.get("/api/commons/allplus", () => {
       return HttpResponse.json(commonsPlusFixtures.threeCommonsPlus, {
+        status: 200,
+      });
+    }),
+  ],
+};
+
+export const adminListPageEmpty = () => <AdminListCommonsPageV2 />;
+
+adminListPageEmpty.parameters = {
+  msw: [
+    http.get("/api/currentUser", () => {
+      return HttpResponse.json(apiCurrentUserFixtures.adminUser, {
+        status: 200,
+      });
+    }),
+    http.get("/api/systemInfo", () => {
+      return HttpResponse.json(systemInfoFixtures.showingNeither, {
+        status: 200,
+      });
+    }),
+    http.get("/api/commons/allplus", () => {
+      return HttpResponse.json([], {
         status: 200,
       });
     }),
